Refactor time calculation and formatting logic
Updated time handling in service classes to improve accuracy and consistency. Key changes include:
- Adjusted total worked hours formatting in `ActivityAppService.cs`.
- Removed static culture info in `ReportAppService.cs` and streamlined methods.
- Enhanced `WorkSummaryAppService.cs` to include unaccounted work duration in response models.
- Added unaccounted work duration property to `TotalWorkHoursResponseViewModel` and `WorkTimeCalculationDto`.
- Introduced `CalcTotalDurationInMinutes` method in `ActivityPeriod`.
- Refined effective work time calculation logic in `ActivityWorkCalculatorService.cs`.
- Updated `FormatHelper` to format numbers as doubles for consistency.

+++ src/dm.PulseShift.Infra.CrossCutting.Shared/Helpers/FormatHelper.cs
@@ -7,7 +7,7 @@ public static class FormatHelper
     public static readonly CultureInfo PtBrCulture = CultureInfo.GetCultureInfo("pt-BR");
     public const string BrazilianDateTimeFormat = "dd/MM/yyyy HH:mm";
 
-    public static string FormatNumberToBrazilianString(decimal number) => number.ToString("F2", PtBrCulture);
+    public static string FormatNumberToBrazilianString(double number) => number.ToString("F2", PtBrCulture);
     public static string FormatDateTimeToBrazilianString(DateTime date, string format = BrazilianDateTimeFormat)
     {
         return date.ToString(format, PtBrCulture);
